Simplify quality control
The distinction between hd/fullhd was pretty invisible for me. And these
constraints are just targets, so no guarantee the camera will use
exactly.

So I'll drop "hd", and relabel the options to be less specific.
Additionally, we set "normal" quality when initing, and this option will
appear selected if the select element is interacted with.